+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr 0.9.0 (in development)
 
+* `mutate()` and `summarise()` can now modify grouping variables (#4709).
+
 * Grouped data frames now have `[[<-`, `[<-` and `$<-` methods that will 
   re-generate the underlying grouping. Note that modifying grouping variables 
   in multiple steps (i.e. `df$grp1 <- 1; df$grp2 <- 1`) will be inefficient
@@ -7,6 +9,7 @@
 
 * `[.grouped_df` now regroups to respect any grouping columns that have
   been removed (#4708).
+
 * `as.tbl()` and `tbl_df()` have been formally deprecated.
   Please use `as_tibble()` instead.